fix stack errors
seems like the connections list is part of the stack struct, not of the toggle switch itself (despite it usually being the last element in a stack).
So once you try adding multiple toggle switches, there were problems.
I seem to have been previously confused by this system,
as seen with the _unknown1 field which was just an empty connections list.